+++ R/mvnormal_normal_wishart.R
@@ -4,12 +4,14 @@
 #' @export
 MvnormalCreate <- function(priorParameters) {
 
-  mdobj <- MixingDistribution("mvnormal", priorParameters, "conjugate")
+  mdObj <- MixingDistribution("mvnormal", priorParameters, "conjugate")
 
-  return(mdobj)
+  return(mdObj)
 }
 
-Likelihood.mvnormal <- function(mdobj, x, theta) {
+#' @export
+#' @rdname Likelihood
+Likelihood.mvnormal <- function(mdObj, x, theta) {
 
   y <- sapply(seq_len(dim(theta[[1]])[3]),
             function(i) mvtnorm::dmvnorm(x, theta[[1]][,, i], theta[[2]][, , i]))
@@ -17,9 +19,11 @@ Likelihood.mvnormal <- function(mdobj, x, theta) {
   return(y)
 }
 
-PriorDraw.mvnormal <- function(mdobj, n = 1) {
+#' @export
+#' @rdname PriorDraw
+PriorDraw.mvnormal <- function(mdObj, n = 1) {
 
-  priorParameters <- mdobj$priorParameters
+  priorParameters <- mdObj$priorParameters
 
   sig <- rWishart(n, priorParameters$nu, priorParameters$Lambda)
 
@@ -30,9 +34,11 @@ PriorDraw.mvnormal <- function(mdobj, n = 1) {
   return(theta)
 }
 
-PosteriorDraw.mvnormal <- function(mdobj, x, n = 1) {
+#' @export
+#' @rdname PosteriorDraw
+PosteriorDraw.mvnormal <- function(mdObj, x, n = 1, ...) {
 
-  post_parameters <- PosteriorParameters(mdobj, x)
+  post_parameters <- PosteriorParameters(mdObj, x)
 
   sig <- rWishart(n, post_parameters$nu_n, post_parameters$t_n)
   mu <- simplify2array(lapply(seq_len(n),
@@ -43,17 +49,19 @@ PosteriorDraw.mvnormal <- function(mdobj, x, n = 1) {
   return(list(mu = mu, sig = sig/post_parameters$kappa_n^2))
 }
 
-PosteriorParameters.mvnormal <- function(mdobj, x) {
+#' @export
+#' @rdname PosteriorParameters
+PosteriorParameters.mvnormal <- function(mdObj, x) {
 
   if (!is.matrix(x)) {
     x <- matrix(x, ncol = length(x))
   }
 
-  kappa0 <- mdobj$priorParameters$kappa0
-  mu0 <- mdobj$priorParameters$mu0
+  kappa0 <- mdObj$priorParameters$kappa0
+  mu0 <- mdObj$priorParameters$mu0
 
   kappa_n <- kappa0 + nrow(x)
-  nu_n <- mdobj$priorParameters$nu + nrow(x)
+  nu_n <- mdObj$priorParameters$nu + nrow(x)
 
   mu_n <- (kappa0 * mu0 + nrow(x) * colMeans(x))/(nrow(x) + kappa0)
 
@@ -61,22 +69,24 @@ PosteriorParameters.mvnormal <- function(mdobj, x) {
 
   sum_squares[is.na(sum_squares)] <- 0
 
-  t_n <- mdobj$priorParameters$Lambda + sum_squares + ((kappa0 * nrow(x))/(kappa0 +
+  t_n <- mdObj$priorParameters$Lambda + sum_squares + ((kappa0 * nrow(x))/(kappa0 +
     nrow(x))) * ((mu0 - colMeans(x)) %*% t(mu0 - colMeans(x)))
 
   return(list(mu_n = mu_n, t_n = t_n, kappa_n = kappa_n, nu_n = nu_n))
 }
 
-Predictive.mvnormal <- function(mdobj, x) {
+#' @export
+#' @rdname Predictive
+Predictive.mvnormal <- function(mdObj, x) {
 
-  priorParameters <- mdobj$priorParameters
+  priorParameters <- mdObj$priorParameters
   pred <- numeric(nrow(x))
 
   d <- ncol(x)
 
   for (i in seq_along(pred)) {
 
-    post_params <- PosteriorParameters(mdobj, x[i, ,drop=FALSE])
+    post_params <- PosteriorParameters(mdObj, x[i, ,drop=FALSE])
 
     pred[i] <- (pi^(-nrow(x[i,,drop=FALSE]) * d/2))
     pred[i] <- pred[i] * (priorParameters$kappa0/post_params$kappa_n)^(d/2)