+++ tests/testthat/test_utilities.R
@@ -11,4 +11,16 @@ test_that("Weighted Function Generator", {
 
   expect_is(test_weighted_func, "function")
   expect_equal(test_weighted_func(0), dnorm(0))
+  expect_equivalent(test_weighted_func(as.matrix(0)), dnorm(0))
+  expect_equivalent(test_weighted_func(as.data.frame(0)), dnorm(0))
+})
+
+
+
+test_that("Burn", {
+  dp <- Fit(DirichletProcessGaussian(rnorm(10)), 4)
+
+  burned_dp <- Burn(dp, 2)
+
+  expect_s3_class(burned_dp, "dirichletprocess")
 })